Change: Rename Tooltip module to ToolTip
Use same PascalCase pattern as the component itself.
