+++ src/web/components/chart/base/Legend.tsx
@@ -9,7 +9,7 @@ import {type ToString} from 'gmp/types';
 import {isDefined} from 'gmp/utils/identity';
 import LegendLabel from 'web/components/chart/base/LagendLabel';
 import {DEFAULT_SHAPE_SIZE} from 'web/components/chart/base/LegendLine';
-import ToolTip, {type ToolTipRef} from 'web/components/chart/base/Tooltip';
+import ToolTip, {type ToolTipRef} from 'web/components/chart/base/ToolTip';
 import Theme from 'web/utils/theme';
 
 interface RectProps {
