+++ src/web/components/chart/donut/Arc2d.tsx
@@ -6,7 +6,7 @@
 import {isDefined} from 'gmp/utils/identity';
 import Group from 'web/components/chart/base/Group';
 import {type LegendData} from 'web/components/chart/base/Legend';
-import ToolTip from 'web/components/chart/base/Tooltip';
+import ToolTip from 'web/components/chart/base/ToolTip';
 import {type Path} from 'web/components/chart/utils/Path';
 import Theme from 'web/utils/theme';
